Update Frontend Component Steps

+++ roboshop/components/common.sh
@@ -59,8 +59,10 @@ DOWNLOAD() {
   cd /tmp
   unzip -o /tmp/${1}.zip &>>${LOG_FILE}
   STAT_CHECK $? "Extract ${1} Code"
-  rm -rf /home/roboshop/${component} && mkdir -p /home/roboshop/${component} && cp -r /tmp/${component}-main/* /home/roboshop/${component} &>>${LOG_FILE}
-  STAT_CHECK $? "Copy ${component} Content"
+  if [ ! -z "${component}" ]; then
+    rm -rf /home/roboshop/${component} && mkdir -p /home/roboshop/${component} && cp -r /tmp/${component}-main/* /home/roboshop/${component} &>>${LOG_FILE}
+    STAT_CHECK $? "Copy ${component} Content"
+  fi
 }
 
 NODEJS() {
